requirements.txt
kubernetes==25.3.0  # 完全适配k8s v1.24，更多查看https://github.com/kubernetes-client/python?tab=readme-ov-file
sqlmodel>=0.0.34
python-dotenv>=1.2.1
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # passlib 1.7.4 与 bcrypt 4.1.0+ 不兼容，需要使用 4.0.1

fastapi>=0.128.8
uvicorn>=0.39.0
httpx>=0.28.1
hepai>=1.1.40
